Comparison function added
Hey guys,
I’ve added the comparison function to the program.
Also I simulated the market return in a dumb way just to see if it
works…
I have set the market return to be the current stock price*100 plus a
random variable that uniformly distributed over 400 to 600, just to
approximate our initial total amount of money(1000 cash + 50 stocks *
10 as the stock price).
And like the way we’ve discussed before, I tried to use AR/MR -1 to
estimate. It worked. I was just thinking of another way to see the
profitability more intuitively, that is looking at the percentage ratio
of AR/MR. So I wrote it that way in this version.
Besides, the way of calculating AR and MR might also be a little
clumsy, you guys can also modify it if you figured out a better way to
calculate the rate of return.